switch CitiesMap to ts
