+++ package.json
@@ -47,6 +47,7 @@
   },
   "dependencies": {
     "@types/leaflet": "1.5.17",
+    "@types/react-redux": "7.1.9",
     "axios": "0.19.2",
     "leaflet": "1.6.0",
     "prop-types": "15.7.2",
