+++ tsconfig.json
@@ -5,6 +5,7 @@
     "target": "es2016",
     "jsx": "react",
     "allowSyntheticDefaultImports": true,
+    "allowJs": true
   },
   "include": [
     "src/**/*"
